fix(ai): stop CalculateCost fusing multiply-adds (review of upstream 1a584a7a5 port)
ai.CalculateCost differed from pi's calculateCost in the last bit on arm64.
The Go spec lets the compiler fuse x*y + z into one FMA, across statements,
and arm64 does: FMADDD at the cacheWrite sum (short rate * tokens + 2x input
* 1h tokens) and three in the total, which re-read the unrounded bucket
products. V8 rounds every multiply and add separately. usage.cost is
persisted in sessions and sent over the protocol, so the last bit is visible.
The bug predates this slice; the 0.87.1 regen routes new tiered prices through
it (the xai tiers from 1a584a7a5, the gpt-6 sol/luna tiers from db91e0331).
amd64 at GOAMD64=v3 fuses the same way; baseline amd64 does not.

Each product is now converted explicitly, float64(x*y), before it is summed,
which the spec says forbids the fusion. go tool objdump of CalculateCost shows
no FMADDD on arm64, and -S shows no VFMADD at GOAMD64=v3.

TestCalculateCostMatchesPi compares exact bits against 120 rows captured from
the published pi-ai 0.87.1 build by the new testdata/cost/capture.mjs
(dist/models.js calculateCost over dist/models.generated.js rates): every
catalog model with a pricing tier at its threshold and one token above it,
with 1h cache writes; base-rate models from four providers; grok-4.7 on both
sides of its 200k tier; and four usages where cacheWrite differs. Each row
carries its rates, so a regen that moves a price does not stale the file. The
existing TestCalculateCostAppliesInputTiers compares within 1e-12 and could
not see this.

Red, observed by assertion on this darwin/arm64 host before the fix: 22 rows
differ in total and 4 in cacheWrite, e.g.
  xai/grok-4.7 (input 164300, ..., cacheWrite1h 167): cost.total = 0.374031,
    pi 0.37403100000000006
  xai/grok-4.7 (input 184451, ..., cacheWrite1h 918): cost.total =
    0.7795870000000001, pi 0.779587
  openai-codex/gpt-6-luna (input 265984, ..., cacheWrite1h 299):
    cost.cacheWrite = 7.042499999999999e-05, pi 7.042500000000002e-05
Green after. A scratch sweep outside the commit (every catalog model, 6070
usages, and 400 more grok-4.7 usages around the tier) went from 1024 differing
rows to 0.

+++ ai/models.go
@@ -69,12 +69,18 @@ func CalculateCost(model *Model, usage *Usage) CostBreakdown {
 	}
 
 	// Anthropic charges 2x base input for 1h cache writes.
+	//
+	// Each product is converted explicitly before it is summed. The Go spec
+	// lets the compiler fuse x*y + z into one FMA (arm64 does; so does amd64 at
+	// GOAMD64=v3), rounding once where V8 rounds the multiply and the add
+	// separately, and the conversion is what forbids that fusion. Without it the
+	// total and cacheWrite land a last bit off pi's on some usages.
 	longWrite := usage.CacheWrite1h
 	shortWrite := usage.CacheWrite - longWrite
-	usage.Cost.Input = rateInput / 1_000_000 * float64(usage.Input)
-	usage.Cost.Output = rateOutput / 1_000_000 * float64(usage.Output)
-	usage.Cost.CacheRead = rateCacheRead / 1_000_000 * float64(usage.CacheRead)
-	usage.Cost.CacheWrite = (rateCacheWrite*float64(shortWrite) + rateInput*2*float64(longWrite)) / 1_000_000
+	usage.Cost.Input = float64(rateInput / 1_000_000 * float64(usage.Input))
+	usage.Cost.Output = float64(rateOutput / 1_000_000 * float64(usage.Output))
+	usage.Cost.CacheRead = float64(rateCacheRead / 1_000_000 * float64(usage.CacheRead))
+	usage.Cost.CacheWrite = (float64(rateCacheWrite*float64(shortWrite)) + float64(rateInput*2*float64(longWrite))) / 1_000_000
 	usage.Cost.Total = usage.Cost.Input + usage.Cost.Output + usage.Cost.CacheRead + usage.Cost.CacheWrite
 	return usage.Cost
 }
